Jewelry_RAG/requirements.txt
streamlit>=1.32.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
Pillow>=10.0.0
python-dotenv>=1.0.0
google-genai>=0.8.0
SpeechRecognition>=3.10.0
